chore: backend routers/schemas, docs, frontend router and deps
Made-with: Cursor

+++ backend/models/monitor_remote.py
@@ -10,7 +10,7 @@ class MonitorRemoteConfig(Base):
     __tablename__ = "monitor_remote_config"
 
     id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    user_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True)
+    user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True)
     host: Mapped[str] = mapped_column(String(255), nullable=False)
     port: Mapped[int] = mapped_column(Integer, nullable=False, default=22)
     username: Mapped[str] = mapped_column(String(64), nullable=False)
